Move grain modules into the ledger folder
The grain and grainAllocation modules should be included as part of the
ledger, rather than living in a whole separate subfolder.

Test plan: Just a rename; `yarn test` passes
